+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat(product_enablement): Add enable/disable support for API Discovery ([#1111](https://github.com/fastly/terraform-provider-fastly/pull/1111))
 - feat(domainsv1/data source): add support for the v1 domains data source ([#1112](https://github.com/fastly/terraform-provider-fastly/pull/1112))
 - feat(service): 'domain' blocks are now optional ([#1113](https://github.com/fastly/terraform-provider-fastly/pull/1113))
+- feat(domain_service_link): add support for domain service links ([#1110](https://github.com/fastly/terraform-provider-fastly/pull/1110))
 
 ### BUG FIXES: